requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
jinja2==3.1.6
    # via pktmask (pyproject.toml)
markdown==3.8
    # via pktmask (pyproject.toml)
markupsafe==3.0.2
    # via
    #   jinja2
    #   pktmask (pyproject.toml)
packaging==25.0
    # via pktmask (pyproject.toml)
pydantic==2.11.5
    # via pktmask (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
pyqt6==6.9.0
    # via pktmask (pyproject.toml)
pyqt6-qt6==6.9.0
    # via
    #   pktmask (pyproject.toml)
    #   pyqt6
pyqt6-sip==13.10.2
    # via
    #   pktmask (pyproject.toml)
    #   pyqt6
pyyaml==6.0.2
    # via pktmask (pyproject.toml)
scapy==2.6.1
    # via pktmask (pyproject.toml)
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic

# The following packages are considered to be unsafe in a requirements file:
# setuptools
